Build the set() method on HashMap
For that, I almost asked for help on the discord because I didn't
understand if I needed to change my linked list methods for it to work
with an array of [key, value], or if I missed something on assignment.
I eventually jsut rewrote some special methods for containsKey() and
findKey(), which took me like 10min.