Finish the project
I finished the entries method and growing automatically the array when
the limit is reached. Indeed you needed to finish all of the other
methods to be able to implement the functionnality to expand twice it's
size. Because I didn't think I would need to reset every pairs after
expanding, but it's necessary because the hashing function changes and
so you can't access the old hashed pairs after expended. But no sweat !